+++ http-utils/src/main/java/ai/vespa/util/http/hc4/retry/RetryConsumer.java
@@ -1,5 +1,5 @@
 // Copyright 2020 Oath Inc. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-package ai.vespa.util.http.retry;
+package ai.vespa.util.http.hc4.retry;
 
 import org.apache.http.client.protocol.HttpClientContext;
