+++ http-utils/src/test/java/ai/vespa/util/http/hc4/retry/DelayedResponseLevelRetryHandlerTest.java
@@ -1,5 +1,5 @@
 // Copyright Verizon Media. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-package ai.vespa.util.http.retry;
+package ai.vespa.util.http.hc4.retry;
 
 import org.apache.http.HttpResponse;
 import org.apache.http.HttpStatus;
